refactor: splits and moves checkout context
